:hammer: Refactor icons to typescript. #120
