+++ .eslintrc
@@ -65,7 +65,7 @@
     "no-extra-bind": ["error"],
     "no-extra-boolean-cast": ["error"],
     "no-inline-comments": ["error"],
-    "no-implicit-coercion": ["error"],
+    "no-implicit-coercion": ["off"],
     "no-implied-eval": ["error"],
     "no-inner-declarations": ["off"],
     "no-invalid-this": ["error"],
