+++ package.json
@@ -21,7 +21,7 @@
     "build:dazzler::calendar": "dazzler generate src/calendar/js/components dazzler/components/calendar --ts",
     "build:dazzler::markdown": "dazzler generate src/markdown/js/components dazzler/components/markdown",
     "build:dazzler::auth": "dazzler generate src/auth/js/components dazzler/components/auth --ts",
-    "build:dazzler::icons": "dazzler generate src/icons/js/components dazzler/components/icons",
+    "build:dazzler::icons": "dazzler generate src/icons/ts/components dazzler/components/icons --ts",
     "build:dazzler::electron": "dazzler generate src/electron/renderer/components dazzler/components/electron --ts",
     "build:dazzler::watch": "run-p build:dazzler::core build:dazzler::extra build:dazzler::electron",
     "build:dazzler": "run-p build:dazzler::core build:dazzler::test build:dazzler::extra build:dazzler::calendar build:dazzler::markdown build:dazzler::auth build:dazzler::icons build:dazzler::ts build:dazzler::electron build:dazzler::html build:dazzler::svg",
