--- src/icons/js/IconContext.js
@@ -1,6 +1,0 @@
-import React from 'react';
-export default React.createContext({
-    packs: [],
-    // eslint-disable-next-line @typescript-eslint/no-empty-function
-    addPack: () => {},
-});
