+++ src/icons/ts/IconContext.ts
@@ -1,0 +1,6 @@
+import React from 'react';
+export default React.createContext<IconContextType>({
+    packs: {},
+    addPack: () => null,
+    isLoaded: () => false,
+});
