+++ src/icons/ts/components/TypiconsPack.tsx
@@ -1,18 +1,18 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import IconPack from './IconPack';
+import {DazzlerProps} from '../../../commons/js/types';
 
 /**
  * Free Icon pack from: https://www.s-ings.com/typicons/
  *
- * Pack name: ``typcn``
+ * :Pack: ``typcn``
  *
  * @example
  *
  *     icons.TypiconsPack()
  *     icons.Icon('typcn-globe')
  */
-const TypiconsPack = () => {
+const TypiconsPack = (_: DazzlerProps) => {
     return (
         <IconPack
             name="typcn"
@@ -22,12 +22,5 @@ const TypiconsPack = () => {
 };
 
 TypiconsPack.defaultProps = {};
-
-TypiconsPack.propTypes = {
-    class_name: PropTypes.string,
-    style: PropTypes.object,
-    identity: PropTypes.string,
-    updateAspects: PropTypes.func,
-};
 
 export default TypiconsPack;
