+++ src/icons/ts/types.d.ts
@@ -1,0 +1,14 @@
+type IconPackDict = {
+    [name: string]: IconPackType;
+};
+
+type IconPackType = {
+    url: string;
+    name: string;
+}
+
+type IconContextType = {
+    packs: IconPackDict;
+    addPack: (pack: IconPackType) => void;
+    isLoaded: (packName: string) => boolean;
+}
